Issue #763 Review adjustments
This commit removes the ServerStoreProxyImpl class, changes
ServerStoreProxy to a class and moves most of the former content
of ServerStoreProxyImpl into ServerStoreProxy.  It also removes
some of the ServerStoreProxy bookkeeping from DefaultClusteringService
and drops the 'create' method from PersistableResourceService.
