+++ clustered/build.gradle
@@ -15,11 +15,13 @@
  */
 
 ext {
-    coordinatorVersion = '1.0-SNAPSHOT'
-    offheapResourceVersion = '1.0-SNAPSHOT'
-    entityApiVersion = '1.0-SNAPSHOT'
-    entityTestLibVersion = '1.0-SNAPSHOT'
+    terracottaPlatformVersion = '1.0-SNAPSHOT'
+    terracottaApisVersion = '1.0-SNAPSHOT'
     terracottaCoreVersion = '5.0.0-SNAPSHOT'
+    coordinatorVersion = terracottaPlatformVersion
+    offheapResourceVersion = terracottaPlatformVersion
+    entityApiVersion = terracottaApisVersion
+    entityTestLibVersion = terracottaApisVersion
 }
 
 subprojects {
