+++ clustered/client/src/main/java/org/ehcache/clustered/client/service/ClusteringService.java
@@ -41,16 +41,14 @@ public interface ClusteringService extends PersistableResourceService {
    *
    * @return a new {@code ServerStoreProxy}
    */
-  <K, V> ServerStoreProxy<K, V> getServerStoreProxy(ClusteredCacheIdentifier cacheIdentifier, final Store.Configuration<K, V> storeConfig);
+  <K, V> ServerStoreProxy getServerStoreProxy(ClusteredCacheIdentifier cacheIdentifier, final Store.Configuration<K, V> storeConfig);
 
   /**
    * Releases access to a {@link ServerStoreProxy} and the server-resident {@code ServerStore} it represents.
    *
    * @param storeProxy a {@code ServerStoreProxy} obtained through {@link #getServerStoreProxy}
-   * @param <K> the cache-expose key type
-   * @param <V> the cache-expose value type
    */
-  <K, V> void releaseServerStoreProxy(ServerStoreProxy<K, V> storeProxy);
+  void releaseServerStoreProxy(ServerStoreProxy storeProxy);
 
   void connect();
